Add existing project http client creation process
