+++ packages/ballerina-extension/e2e-test/e2e-playwright-tests/data/http_try_it_existing_project/.vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "ballerina.isBI": true,
+    "files.saveConflictResolution": "overwriteFileOnDisk",
+    "files.hotExit": "off",
+    "files.autoSave": "off"
+}
