+++ packages/ballerina-extension/e2e-test/e2e-playwright-tests/data/http_try_it_existing_project/functions.bal
@@ -1,0 +1,1 @@
+
