+++ packages/ballerina-extension/e2e-test/e2e-playwright-tests/data/http_try_it_existing_project/service.bal
@@ -1,0 +1,13 @@
+import ballerina/http;
+
+listener http:Listener httpListener = new (9090);
+
+// Pre-baked HTTP service (per the e2e-writer rule that scenarios must not
+// modify Ballerina sources at runtime): already-defined service + resource so
+// the test can open the project and go straight to Try It, without an
+// authoring step to create it.
+service / on httpListener {
+    resource function get greeting() returns json {
+        return {message: "Hello, Ballerina!"};
+    }
+}
